requirements.txt
feedparser>=6.0,<7.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
beautifulsoup4>=4.12,<5.0
lxml>=5.0,<6.0
